+++ gcal_routine/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 2.8.3)
-project(ical_routine)
+project(gcal_routine)
 
 ## Find catkin macros and libraries
 ## if COMPONENTS list like find_package(catkin REQUIRED COMPONENTS xyz)
@@ -82,7 +82,7 @@ catkin_python_setup()
 ## DEPENDS: system dependencies of this project that dependent projects also need
 catkin_package(
 #  INCLUDE_DIRS include
-#  LIBRARIES ical_routine
+#  LIBRARIES gcal_routine
 #  CATKIN_DEPENDS message_generation rospy std_msgs strands_executive_msgs
 #  DEPENDS system_lib
 )
@@ -99,19 +99,19 @@ include_directories(
 )
 
 ## Declare a cpp library
-# add_library(ical_routine
-#   src/${PROJECT_NAME}/ical_routine.cpp
+# add_library(gcal_routine
+#   src/${PROJECT_NAME}/gcal_routine.cpp
 # )
 
 ## Declare a cpp executable
-# add_executable(ical_routine_node src/ical_routine_node.cpp)
+# add_executable(gcal_routine_node src/gcal_routine_node.cpp)
 
 ## Add cmake target dependencies of the executable/library
 ## as an example, message headers may need to be generated before nodes
-# add_dependencies(ical_routine_node ical_routine_generate_messages_cpp)
+# add_dependencies(gcal_routine_node gcal_routine_generate_messages_cpp)
 
 ## Specify libraries to link a library or executable target against
-# target_link_libraries(ical_routine_node
+# target_link_libraries(gcal_routine_node
 #   ${catkin_LIBRARIES}
 # )
 
@@ -130,7 +130,7 @@ include_directories(
 # )
 
 ## Mark executables and/or libraries for installation
-# install(TARGETS ical_routine ical_routine_node
+# install(TARGETS gcal_routine gcal_routine_node
 #   ARCHIVE DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
 #   LIBRARY DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
 #   RUNTIME DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
@@ -155,7 +155,7 @@ include_directories(
 #############
 
 ## Add gtest based cpp test target and link libraries
-# catkin_add_gtest(${PROJECT_NAME}-test test/test_ical_routine.cpp)
+# catkin_add_gtest(${PROJECT_NAME}-test test/test_gcal_routine.cpp)
 # if(TARGET ${PROJECT_NAME}-test)
 #   target_link_libraries(${PROJECT_NAME}-test ${PROJECT_NAME})
 # endif()
