+++ gcal_routine/package.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0"?>
 <package>
-  <name>ical_routine</name>
+  <name>gcal_routine</name>
   <version>0.0.0</version>
-  <description>The ical_routine package</description>
+  <description>The gcal_routine package</description>
 
   <!-- One maintainer tag required, multiple allowed, one person per tag --> 
   <!-- Example:  -->
@@ -19,7 +19,7 @@
   <!-- Url tags are optional, but mutiple are allowed, one per tag -->
   <!-- Optional attribute type can be: website, bugtracker, or repository -->
   <!-- Example: -->
-  <!-- <url type="website">http://wiki.ros.org/ical_routine</url> -->
+  <!-- <url type="website">http://wiki.ros.org/gcal_routine</url> -->
 
 
   <!-- Author tags are optional, mutiple are allowed, one per tag -->
