+++ gcal_routine/setup.py
@@ -3,7 +3,7 @@ from distutils.core import setup
 from catkin_pkg.python_setup import generate_distutils_setup
 
 d = generate_distutils_setup(
-    packages=['ical_routine'],
+    packages=['gcal_routine'],
     package_dir={'': 'src'}
 )
