+++ gcal_routine/tests/gcal_tester.test
@@ -1,0 +1,5 @@
+<launch>
+
+  <test test-name="gcal_tester" pkg="gcal_routine" type="gcal_tester.py" />
+
+</launch>
